# perf: improve compute greatest element method