**Old method**

╔═══════════════╤═════════╤═══════════════╤═══════════╗
║ Slower tests  │ Samples │        Result │ Tolerance ║
╟───────────────┼─────────┼───────────────┼───────────╢
║ Fastest test  │ Samples │        Result │ Tolerance ║
╟───────────────┼─────────┼───────────────┼───────────╢
║ 10_000 events │    3000 │ 169.58 op/sec │  ± 0.63 % ║
╚═══════════════╧═════════╧═══════════════╧═══════════╝

╔══════════════╤═════════╤═══════════════╤═══════════╗
║ Slower tests │ Samples │        Result │ Tolerance ║
╟──────────────┼─────────┼───────────────┼───────────╢
║ Fastest test │ Samples │        Result │ Tolerance ║
╟──────────────┼─────────┼───────────────┼───────────╢
║ all events   │    3000 │ 169.38 op/sec │  ± 0.63 % ║
╚══════════════╧═════════╧═══════════════╧═══════════╝


**New method**
╔═══════════════╤═════════╤═══════════════╤═══════════╗
║ Slower tests  │ Samples │        Result │ Tolerance ║
╟───────────────┼─────────┼───────────────┼───────────╢
║ Fastest test  │ Samples │        Result │ Tolerance ║
╟───────────────┼─────────┼───────────────┼───────────╢
║ 10_000 events │    3000 │ 223.77 op/sec │  ± 0.49 % ║
╚═══════════════╧═════════╧═══════════════╧═══════════╝

╔══════════════╤═════════╤═══════════════╤═══════════╗
║ Slower tests │ Samples │        Result │ Tolerance ║
╟──────────────┼─────────┼───────────────┼───────────╢
║ Fastest test │ Samples │        Result │ Tolerance ║
╟──────────────┼─────────┼───────────────┼───────────╢
║ all events   │    3000 │ 225.22 op/sec │  ± 0.48 % ║
╚══════════════╧═════════╧═══════════════╧═══════════╝